Split coordinator capabilities and write/schedule logic into mixins
