+++ .github/workflows/ci.yaml
@@ -40,7 +40,10 @@ jobs:
           python -m pip install -e ".[dev]"
           python -m pip install homeassistant==${{ matrix.ha-version }} hacs
 
-      - name: Ruff
+      - name: Ruff (custom_components gate)
+        run: ruff check custom_components/
+
+      - name: Ruff (repo)
         run: ruff check .
 
       - name: Black
