+++ custom_components/thessla_green_modbus/diagnostics.py
@@ -5,8 +5,8 @@ Includes the same translated error and status codes as exposed by the ``error_co
 
 from __future__ import annotations
 
-import copy
 import asyncio
+import copy
 import ipaddress
 import logging
 import re
@@ -133,9 +133,16 @@ async def async_get_config_entry_diagnostics(
         translations = await translation.async_get_translations(
             hass, hass.config.language, f"component.{DOMAIN}"
         )
-    except (OSError, ValueError, HomeAssistantError, RuntimeError) as err:  # pragma: no cover - defensive
+    except (
+        OSError,
+        ValueError,
+        HomeAssistantError,
+        RuntimeError,
+    ) as err:  # pragma: no cover - defensive
         _LOGGER.debug("Translation load failed: %s", err)
-    except BaseException as err:  # pragma: no cover - defensive for unexpected translation-layer errors
+    except (
+        BaseException
+    ) as err:  # pragma: no cover - defensive for unexpected translation-layer errors
         if isinstance(err, (KeyboardInterrupt, SystemExit, asyncio.CancelledError)):
             raise
         _LOGGER.debug("Translation load failed unexpectedly: %s", err)
