+++ custom_components/thessla_green_modbus/modbus_helpers.py
@@ -101,13 +101,25 @@ async def async_maybe_await_close(obj: Any | None) -> None:
 
     try:
         result = close()
-    except (AttributeError, OSError, RuntimeError, TypeError, ValueError) as exc:  # pragma: no cover - defensive
+    except (
+        AttributeError,
+        OSError,
+        RuntimeError,
+        TypeError,
+        ValueError,
+    ) as exc:  # pragma: no cover - defensive
         _LOGGER.debug("Error calling close on Modbus client: %s", exc)
         return
 
     try:
         await async_maybe_await(result)
-    except (AttributeError, OSError, RuntimeError, TypeError, ValueError) as exc:  # pragma: no cover - defensive
+    except (
+        AttributeError,
+        OSError,
+        RuntimeError,
+        TypeError,
+        ValueError,
+    ) as exc:  # pragma: no cover - defensive
         _LOGGER.debug("Error awaiting Modbus client close: %s", exc)
 
 
@@ -350,9 +362,7 @@ async def _call_modbus(
         ValueError,
     ) as err:
         if isinstance(err, ModbusIOException) and "request cancelled" in str(err).lower():
-            _LOGGER.debug(
-                "Call to %s cancelled on attempt %s/%s", func_name, attempt, max_attempts
-            )
+            _LOGGER.debug("Call to %s cancelled on attempt %s/%s", func_name, attempt, max_attempts)
         else:
             _LOGGER.debug("Call to %s failed on attempt %s/%s", func_name, attempt, max_attempts)
         raise
@@ -363,7 +373,7 @@ async def _call_modbus(
         except (AttributeError, ValueError, TypeError, UnicodeError) as err:
             _LOGGER.debug("Failed to encode Modbus response: %s", err)
             encoded = b""
-        except (AttributeError, OSError, RuntimeError, TypeError, ValueError) as err:  # pragma: no cover - unexpected
+        except (OSError, RuntimeError) as err:  # pragma: no cover - unexpected
             _LOGGER.exception("Unexpected error encoding Modbus response: %s", err)
             encoded = b""
         if encoded:
