+++ custom_components/thessla_green_modbus/scanner_helpers.py
@@ -102,10 +102,10 @@ SAFE_REGISTERS: list[tuple[int, str]] = [
 ]
 
 __all__ = [
+    "MAX_BATCH_REGISTERS",
     "REGISTER_ALLOWED_VALUES",
+    "SAFE_REGISTERS",
     "SETTING_PREFIX",
     "SPECIAL_VALUE_DECODERS",
-    "MAX_BATCH_REGISTERS",
     "UART_OPTIONAL_REGS",
-    "SAFE_REGISTERS",
 ]
